Add option to overwrite all tags.
In my opinion this is strictly better than the current code that semi-replaces tags but also kinda doesn't. This way no out-of-date info is left hanging around and if you're using tags for this generated info then it's basically impossible to use them for their original purpose anyway.

However, I want to consider and test things more before I make this the default option.