Add mangalist template imports
Wasn't as hard as I thought. Knock on wood.

Additionally, blank CSS no longer leave the START/END marks in the code.

KNOWN ISSUES:
If both classic styles are blank it will only update the first one, which may not be the correct one. I can think of several ways to fix this, but am unsure which would be best or how much it actually matters.